--- .github/actions/update_version/action.yml
@@ -1,13 +1,0 @@
-name : "Update version"
-description : "Updates version based on input ['patch', 'minor' or  'major']"
-inputs:
-  types:
-    description: "Array of all types of updates, type can be 'patch', 'minor' or 'major'"
-    required: false
-    default: 'patch'
-  repo-token:
-    description: "Github repository token with push authorisation"
-    required: true
-runs:
-  using: 'node16'
-  main: updateVersion.js
